* release: (3280 commits)
  Update: Grunt
  Update: Grunt string-replace
  Update: Version name update
  Added change log
  Pulled api changes from repo
  grunt
  PROD-9102 - make the text translatable
  PROD-9102 - make the text translatable
  PROD-9102 - removed error log
  PROD-9102 - phpcbf fixes
  PROD-9102 - Critical Security Vulnerability Fixed - Document Download Function
  ARCH-166 - removed beta tag for readylaunch navigation menu
  ARCH-166 - updated links for the integrations tab
  PROD-8728 - Added date length validation for search term
  PROD-8728 - Fixed elipsed serach issue with other languages
  claude auto review changes
  [ARCH-166] Match Onboarding steps content with Figma
  Revert "Removed synchronize"
  PROD-9024 - use in_array to check strict comparison
  [PROD-9052] PROD-8946 - Implicitly marking parameter as nullable is deprecated; the explicit nullable type must be used instead
  ...

+++ src/bp-performance/classes/integrations/class-bb-members.php
@@ -224,6 +224,8 @@ class BB_Members extends Integration_Abstract {
 	 * @param BP_Activity_Follow $follow Object of Follow.
 	 */
 	public function event_bp_start_following( $follow ) {
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bbapp-deeplinking' );
 		$this->purge_item_cache_by_item_id( $follow->leader_id );
 	}
 
@@ -233,6 +235,8 @@ class BB_Members extends Integration_Abstract {
 	 * @param BP_Activity_Follow $follow Object of Follow.
 	 */
 	public function event_bp_follow_start_following( $follow ) {
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bbapp-deeplinking' );
 		$this->purge_item_cache_by_item_id( $follow->leader_id );
 	}
 
@@ -242,6 +246,8 @@ class BB_Members extends Integration_Abstract {
 	 * @param BP_Activity_Follow $follow Object of Follow.
 	 */
 	public function event_bp_stop_following( $follow ) {
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bbapp-deeplinking' );
 		$this->purge_item_cache_by_item_id( $follow->leader_id );
 	}
 
@@ -251,6 +257,8 @@ class BB_Members extends Integration_Abstract {
 	 * @param BP_Activity_Follow $follow Object of Follow.
 	 */
 	public function event_bp_follow_stop_following( $follow ) {
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $follow->follower_id, 'bbapp-deeplinking' );
 		$this->purge_item_cache_by_item_id( $follow->leader_id );
 	}
 
@@ -264,6 +272,10 @@ class BB_Members extends Integration_Abstract {
 	public function event_friends_friendship_requested( $friendship_id, $initiator_user_id, $friend_user_id ) {
 		$this->purge_item_cache_by_item_id( $initiator_user_id );
 		$this->purge_item_cache_by_item_id( $friend_user_id );
+		Cache::instance()->purge_by_user_id( $initiator_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friend_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $initiator_user_id, 'bbapp-deeplinking' );
+		Cache::instance()->purge_by_user_id( $friend_user_id, 'bbapp-deeplinking' );
 	}
 
 	/**
@@ -276,6 +288,10 @@ class BB_Members extends Integration_Abstract {
 	public function event_friends_friendship_accepted( $friendship_id, $initiator_user_id, $friend_user_id ) {
 		$this->purge_item_cache_by_item_id( $initiator_user_id );
 		$this->purge_item_cache_by_item_id( $friend_user_id );
+		Cache::instance()->purge_by_user_id( $initiator_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friend_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $initiator_user_id, 'bbapp-deeplinking' );
+		Cache::instance()->purge_by_user_id( $friend_user_id, 'bbapp-deeplinking' );
 	}
 
 	/**
@@ -288,6 +304,10 @@ class BB_Members extends Integration_Abstract {
 	public function event_friends_friendship_deleted( $friendship_id, $initiator_user_id, $friend_user_id ) {
 		$this->purge_item_cache_by_item_id( $initiator_user_id );
 		$this->purge_item_cache_by_item_id( $friend_user_id );
+		Cache::instance()->purge_by_user_id( $initiator_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friend_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $initiator_user_id, 'bbapp-deeplinking' );
+		Cache::instance()->purge_by_user_id( $friend_user_id, 'bbapp-deeplinking' );
 	}
 
 	/**
@@ -299,6 +319,10 @@ class BB_Members extends Integration_Abstract {
 	public function event_friends_friendship_rejected( $friendship_id, $friendship ) {
 		$this->purge_item_cache_by_item_id( $friendship->initiator_user_id );
 		$this->purge_item_cache_by_item_id( $friendship->friend_user_id );
+		Cache::instance()->purge_by_user_id( $friendship->initiator_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friendship->friend_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friendship->initiator_user_id, 'bbapp-deeplinking' );
+		Cache::instance()->purge_by_user_id( $friendship->friend_user_id, 'bbapp-deeplinking' );
 	}
 
 	/**
@@ -310,6 +334,10 @@ class BB_Members extends Integration_Abstract {
 	public function event_friends_friendship_post_delete( $initiator_userid, $friend_userid ) {
 		$this->purge_item_cache_by_item_id( $initiator_userid );
 		$this->purge_item_cache_by_item_id( $friend_userid );
+		Cache::instance()->purge_by_user_id( $initiator_userid, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friend_userid, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $initiator_userid, 'bbapp-deeplinking' );
+		Cache::instance()->purge_by_user_id( $friend_userid, 'bbapp-deeplinking' );
 	}
 
 	/**
@@ -321,6 +349,10 @@ class BB_Members extends Integration_Abstract {
 	public function event_friends_friendship_withdrawn( $friendship_id, $friendship ) {
 		$this->purge_item_cache_by_item_id( $friendship->initiator_user_id );
 		$this->purge_item_cache_by_item_id( $friendship->friend_user_id );
+		Cache::instance()->purge_by_user_id( $friendship->initiator_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friendship->friend_user_id, 'bp-activity' );
+		Cache::instance()->purge_by_user_id( $friendship->initiator_user_id, 'bbapp-deeplinking' );
+		Cache::instance()->purge_by_user_id( $friendship->friend_user_id, 'bbapp-deeplinking' );
 	}
 
 	/**
@@ -359,7 +391,9 @@ class BB_Members extends Integration_Abstract {
 	 */
 	public function event_bp_core_delete_existing_avatar( $args ) {
 		$user_id = ! empty( $args['item_id'] ) ? absint( $args['item_id'] ) : 0;
-		$this->purge_item_cache_by_item_id( $user_id );
+		if ( ! empty( $user_id ) && isset( $args['object'] ) && 'user' === $args['object'] ) {
+			$this->purge_item_cache_by_item_id( $user_id );
+		}
 	}
 
 	/**
@@ -608,6 +642,11 @@ class BB_Members extends Integration_Abstract {
 	 * @param int $user_id User ID.
 	 */
 	private function purge_subscription_cache_by_user_id( $user_id ) {
+
+		if ( empty( $user_id ) ) {
+			return;
+		}
+
 		$args = array(
 			'user_id' => $user_id,
 			'fields'  => 'id',
